fix: downgrade next config for storybook
