+++ next.config.js
@@ -1,0 +1,19 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+/**
+ * @type {import('next').NextConfig}
+ */
+
+const withPWA = require('next-pwa')({
+  dest: 'public',
+  disable: process.env.NODE_ENV !== 'production'
+})
+
+const nextConfig = {
+  reactStrictMode: true,
+  swcMinify: true,
+  compiler: {
+    styledComponents: true
+  }
+}
+
+module.exports = withPWA(nextConfig)
