--- next.config.mjs
@@ -1,21 +1,0 @@
-import nextPWA from 'next-pwa'
-
-const withPWA = nextPWA({
-  dest: 'public',
-  disable: process.env.NODE_ENV !== 'production'
-})
-
-/**
- * @type {import('next').NextConfig}
- */
-
-const nextConfig = {
-  reactStrictMode: true,
-  swcMinify: true,
-  compiler: {
-    styledComponents: true,
-    removeConsole: true
-  }
-}
-
-export default withPWA(nextConfig)
